Fix startup/teardown of qemu vm
This now actually works as expected.
Before, as soon as the stop script exited systemd was killing the qemu
process.

Now this is a bit more fancy.
It relies on sd_notify to actually say the VM is running (via probing
the qemu guest agaent).
It also waits for qemu to exit before the shutdown script exits.