brightnessd: Add restore command to restore previous value
This is useful for dimming the displays when idle and then restoring the
previous brightness when coming back.

Also adds a way to send DDC commands to *all* monitors.
In order to simplify things, the client shell script now requires
specifying the monitor to send commands to, and getting the currently
focussed monitor is extracted to a separate script which gets called in
the hyprland config as an argument to the brightness script.

This change has not been wired up to hypridle yet.